# Fix issues with parsing knockout-3.4.0.js and lodash.min.js