* There was a stackoverflow exception and a timeout due to number->string performance and logic problems
* Now test cases cover the files
* I was also able to unify the double as integer checks and now the positive zero double is cached which is a big win, ECMA still has the negative zero handled as caching does not consider that
* I increased the int and double lookup sizes, added -1 and infinities

## Before

|             Method |   N |        Mean |      Error |     StdDev |     Gen 0 |   Allocated |
|------------------- |---- |------------:|-----------:|-----------:|----------:|------------:|
|              Slice | 100 |    877.4 us |   7.603 us |   7.112 us |  239.2578 |   982.81 KB |
|             Concat | 100 |    991.0 us |   1.550 us |   1.210 us |  257.8125 |   1062.5 KB |
|            Unshift | 100 | 36,026.6 us | 182.054 us | 152.023 us | 7000.0000 | 28785.16 KB |
|               Push | 100 | 23,722.6 us | 439.804 us | 411.393 us | 1625.0000 |   6762.5 KB |
|              Index | 100 | 25,124.1 us |  59.588 us |  49.759 us | 1531.2500 |  6333.59 KB |
|                Map | 100 |  4,644.9 us |  15.578 us |  13.810 us | 1804.6875 |  7420.31 KB |
|              Apply | 100 |  1,040.4 us |   5.970 us |   5.584 us |  267.5781 |  1098.44 KB |
| JsonStringifyParse | 100 |  6,641.6 us |  26.941 us |  25.200 us | 1476.5625 |  6068.75 KB |

| Method |  N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------- |--- |--------:|---------:|---------:|------------:|-----------:|----------:|----------:|
|   ArrayStressBenchmark | 20 | 1.961 s | 0.0741 s | 0.0042 s | 140000.0000 | 12250.0000 | 5750.0000 | 601.17 MB |

|    Method |   N |     Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |---------:|---------:|---------:|------------:|-----------:|----------:|
| UncacheableExpressionsBenchmark| 500 | 572.1 ms | 82.07 ms | 4.637 ms | 114750.0000 | 33562.5000 | 494.73 MB |

## After

|             Method |   N |        Mean |     Error |    StdDev |     Gen 0 |   Allocated |
|------------------- |---- |------------:|----------:|----------:|----------:|------------:|
|              Slice | 100 |    862.7 us |  2.353 us |  1.837 us |  233.3984 |   959.38 KB |
|             Concat | 100 |  1,007.3 us |  1.459 us |  1.364 us |  251.9531 |  1039.06 KB |
|            Unshift | 100 | 33,224.9 us | 32.555 us | 30.452 us | 6937.5000 | 28511.72 KB |
|               Push | 100 | 21,428.9 us | 18.717 us | 17.508 us | 1562.5000 |  6489.06 KB |
|              Index | 100 | 21,517.8 us | 33.752 us | 29.921 us | 1406.2500 |  5814.06 KB |
|                Map | 100 |  4,681.2 us |  8.057 us |  7.537 us | 1804.6875 |  7408.59 KB |
|              Apply | 100 |    992.3 us |  1.772 us |  1.658 us |  263.6719 |  1082.81 KB |
| JsonStringifyParse | 100 |  5,388.7 us | 12.012 us | 10.648 us | 1367.1875 |  5631.25 KB |
  
| Method |  N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |    Gen 2 | Allocated |
|------- |--- |--------:|---------:|---------:|------------:|-----------:|---------:|----------:|
|   ArrayStressBenchmark| 20 | 1.928 s | 0.0283 s | 0.0016 s | 120833.3333 | 16000.0000 | 666.6667 | 524.95 MB |

|    Method |   N |     Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |---------:|---------:|---------:|------------:|-----------:|----------:|
| UncacheableExpressionsBenchmark| 500 | 582.0 ms | 24.04 ms | 1.358 ms | 115520.8333 | 31125.0000 | 494.53 MB |